feat: implement User and Film storages
