+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,0 +1,16 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.List;
+
+public interface FilmStorage {
+
+    Film add(Film film);
+
+    List<Film> getAll();
+
+    Film update(Film film);
+
+    void remove(Film film);
+}
