+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -1,0 +1,16 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.List;
+
+public interface UserStorage {
+
+    User add(User user);
+
+    List<User> getAll();
+
+    User update(User user);
+
+    void remove(User user);
+}
