+++ src/test/java/ru/yandex/practicum/filmorate/controller/FilmControllerTest.java
@@ -7,6 +7,8 @@ import ru.yandex.practicum.filmorate.exception.BadRequestException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
 
 import java.time.LocalDate;
 
@@ -23,7 +25,8 @@ public class FilmControllerTest {
         testFilm.setReleaseDate(LocalDate.of(2000, 10, 10));
         testFilm.setDuration(100);
 
-        testFilmController = new FilmController();
+        FilmStorage testFilmStorage = new InMemoryFilmStorage();
+        testFilmController = new FilmController(testFilmStorage);
     }
 
     @Test
